backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.36
alembic>=1.14.0
pydantic>=2.12.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
pytest>=8.3.4
httpx>=0.28.1
pandas>=2.2.0
pydantic-ai>=1.84.1
annotated-types>=0.7.0